phase 1

+++ anchor/programs/tokenized-yield/src/states/user_stake.rs
@@ -6,7 +6,7 @@ pub struct UserStake {
     pub is_initialized: bool,
     pub owner: Pubkey,
     pub vault: Pubkey,
-    pub quantity: u32,
+    pub quantity: u64,
     pub pay_amount: u64,
     pub debt_claimed: u128,
     pub bump: u8,
